Provide default filters for eBGP usage (#25)
Loop prevention is now provided by the input filter. Prepending the
local ASN and setting the nexthop are provided by the export filter.

Filter overrides now require a single filter to be provided instead of a
list. Custom filters may optionally call the default filter. This change
was made because chaining filters raises correctness and performance
issues. It is better to delegate this complexity to the caller rather
than bake it into the API.