Support structured logging instead of text logging
This greatly increases the user's control over what gets logged and how
it's formatted. Peer events such as session setup and teardown are
logged at INFO and ERROR levels. Session events like BGP announcements
and withdrawals are logged at a locally defined LevelUpdates that falls
between DEBUG and INFO.

This is a breaking change for anyone who enabled logging, but the simple
fix is to just pass slog.Default() instead of log.Default(). The output
will be formatted differently and will not print individual BGP
announcements by default. But that's of course configurable by providing
a custom slog.Handler.