Remove the Network type and move its methods to Table (#49)
This breaking change is necessary for some upcoming efficiency
improvements. We expect most users are calling the Network method on a
Table and then immediately calling AddPath/RemovePath, and for these
cases migrating to the new API is straightforward. If anyone was
iterating over Networks or storing them outside of a table, the
migration will be slightly more involved. We did not find any such use
cases in our internal codebase.